airstrike on compiler warnings
I was sent a log file of a compilation
on FreeBSD with all the nasty cousins
of -Wall enabled.
This is my attempt to fix the majority
of the resulting warnings.

Notable warnings which remain
include the use of zero-length
arrays at the end of structs (struct gmi_set)
and the use of the "%zu" conversion specifier
in printf-like functions to print size_t.

+++ apf/apfAdjReorder.cc
@@ -23,7 +23,7 @@ namespace apf {
     int components = countComponents(field);
     int dim = mesh->getDimension();
     
-    long long dof = 0;
+    long dof = 0;
 
     MeshIterator * it;
     MeshEntity * ent = NULL;
@@ -107,7 +107,7 @@ int AdjReorder(Numbering * num)
 
   // on the off chance there are multiple nodes per vertex...
   int vert_nodes = shape->countNodesOn(Mesh::VERTEX);
-  long long dof_index = dofs-1;
+  long dof_index = dofs-1;
 
   while(dof_index >= 0)
   {
